+++ packages/tools/src/globals.d.ts
@@ -1,3 +1,3 @@
-declare module "htmldiff-js" {
+declare module 'htmldiff-js' {
   export const execute: (a: string, b: string) => string;
 }